backend/requirements.txt
alembic==1.16.5
fastapi==0.116.1
fastapi_mail==1.5.8
uvicorn[standard]==0.35.0

sqlalchemy==2.0.43
pydantic==2.12.4
python-jose[cryptography]==3.5.0
email-validator==2.3.0

redis==7.0.0
httpx==0.28.1

python-dotenv==1.1.1
python-decouple==3.8
python-multipart==0.0.20
pyotp==2.9.0
passlib[argon2]==1.7.4
psycopg2-binary==2.9.11
pypdf==6.1.3


sib-api-v3-sdk==7.6.0

langchain==1.0.7
langchain-core==1.0.5
langchain-google-genai==3.0.3
langchain-community==0.3.29
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-groq==1.0.1
langchain-tavily==0.2.12
langgraph==1.0.3
langchain-text-splitters==0.3.11
faiss-cpu==1.12.0
